增加组件
日历组件：PDTSimpleCalendar
左右滑动列表cell：MGSwipeTableCell

+++ SuperDemo/DMLazyScrollViewViewController.m
@@ -99,7 +99,7 @@
         
         UILabel* label = [[UILabel alloc] initWithFrame:contr.view.bounds];
         label.backgroundColor = [UIColor clearColor];
-        label.text = [NSString stringWithFormat:@"%d",index];
+        label.text = [NSString stringWithFormat:@"%ld",(long)index];
         label.textAlignment = NSTextAlignmentCenter;
         label.font = [UIFont boldSystemFontOfSize:50];
         [contr.view addSubview:label];
